requirements.txt
asgiref==3.6.0
Django==4.1.4
django-jazzmin==2.6.0
djangorestframework==3.14.0
Pillow==9.4.0
psycopg2-binary==2.9.5
python-decouple==3.6
pytz==2022.7.1
sqlparse==0.4.3
